Refactor: Modularize page components and update ContactSection
- Extracted ContactSection from page.jsx to src/app/component/ContactSection.jsx, using the more current inline version.
- Extracted AnimatedBackground, LoadingScreen, Navbar, HeroSection, AboutSection, ServicesSection from page.jsx into separate component files under src/app/component/ui_elements and src/app/component/sections.
- Consolidated ExpertiseSection into src/app/component/ExpertiseSection.jsx using the more detailed version from page.jsx.
- Updated all import paths accordingly in page.jsx and the new component files.
- This improves modularity and maintainability of the main page component while ensuring UI consistency.
