+++ src/app/component/sections/HeroSection.jsx
@@ -1,0 +1,79 @@
+"use client";
+import { motion } from 'framer-motion';
+import { FaReact, FaNode, FaPython, FaRobot } from 'react-icons/fa';
+
+// Hero Section
+const HeroSection = () => {
+  return (
+    <section id="home" className="min-h-screen flex items-center justify-center pt-16">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="text-center">
+          <motion.div
+            initial={{ opacity: 0, y: 30 }}
+            animate={{ opacity: 1, y: 0 }}
+            transition={{ duration: 0.8 }}
+          >
+            <h1 className="text-4xl sm:text-5xl md:text-6xl lg:text-7xl font-bold mb-6">
+              <span className="bg-gradient-to-r from-purple-400 via-pink-400 to-orange-400 bg-clip-text text-transparent">
+                AI-Powered
+              </span>
+              <br />
+              <span className="text-white">Solutions</span>
+            </h1>
+            <p className="text-lg sm:text-xl md:text-2xl text-gray-300 mb-8 max-w-3xl mx-auto">
+              I build intelligent applications that transform businesses through cutting-edge AI integration and modern web technologies.
+            </p>
+          </motion.div>
+          <motion.div
+            className="flex flex-col sm:flex-row gap-4 justify-center items-center mb-12"
+            initial={{ opacity: 0, y: 30 }}
+            animate={{ opacity: 1, y: 0 }}
+            transition={{ duration: 0.8, delay: 0.2 }}
+          >
+            <motion.a
+              href="#contact"
+              className="px-8 py-4 bg-gradient-to-r from-purple-500 to-pink-500 text-white rounded-full text-lg font-semibold hover:shadow-lg hover:shadow-purple-500/30 transition-all"
+              whileHover={{ scale: 1.05 }}
+              whileTap={{ scale: 0.95 }}
+            >
+              Start Your Project
+            </motion.a>
+            <motion.a
+              href="#projects"
+              className="px-8 py-4 border-2 border-purple-500 text-purple-400 rounded-full text-lg font-semibold hover:bg-purple-500/10 transition-all"
+              whileHover={{ scale: 1.05 }}
+              whileTap={{ scale: 0.95 }}
+            >
+              View My Work
+            </motion.a>
+          </motion.div>
+          <motion.div
+            className="flex justify-center items-center space-x-8 text-gray-400"
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            transition={{ duration: 0.8, delay: 0.4 }}
+          >
+            <div className="flex items-center space-x-2">
+              <FaReact className="text-2xl text-cyan-400" />
+              <span>React</span>
+            </div>
+            <div className="flex items-center space-x-2">
+              <FaNode className="text-2xl text-green-400" />
+              <span>Node.js</span>
+            </div>
+            <div className="flex items-center space-x-2">
+              <FaPython className="text-2xl text-yellow-400" />
+              <span>Python</span>
+            </div>
+            <div className="flex items-center space-x-2">
+              <FaRobot className="text-2xl text-purple-400" />
+              <span>AI/ML</span>
+            </div>
+          </motion.div>
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default HeroSection;
