+++ src/app/page.jsx
@@ -11,557 +11,15 @@ import me from './public/images/me.jpeg';
 import { projects } from './projects';
 import { firaCode, spaceGrotesk } from './fonts';
 import SectionTitle from './component/SectionTitle';
+import AnimatedBackground from './component/ui_elements/AnimatedBackground';
+import LoadingScreen from './component/ui_elements/LoadingScreen'; // Kept for now, though EnhancedLoader is used
+import Navbar from './component/navigation/Navbar';
+import HeroSection from './component/sections/HeroSection';
+import AboutSection from './component/sections/AboutSection';
+import ServicesSection from './component/sections/ServicesSection';
+import ExpertiseSection from './component/ExpertiseSection'; // Consolidated to use this
 
-// Animated Background Component
-const AnimatedBackground = () => {
-  const [prefersReducedMotion, setPrefersReducedMotion] = useState(false);
-
-  useEffect(() => {
-    const mediaQuery = window.matchMedia('(prefers-reduced-motion: reduce)');
-    setPrefersReducedMotion(mediaQuery.matches);
-
-    const handleChange = () => setPrefersReducedMotion(mediaQuery.matches);
-    mediaQuery.addEventListener('change', handleChange);
-
-    return () => mediaQuery.removeEventListener('change', handleChange);
-  }, []);
-
-  return (
-    <div className="fixed inset-0 z-[-1] overflow-hidden">
-      {/* Dim overlay for readability */}
-      <div className="absolute inset-0 bg-slate-900/70 z-10 pointer-events-none"></div>
-      {/* Base color */}
-      <div className="absolute inset-0 bg-gradient-to-b from-slate-900 to-slate-800"></div>
-
-      {/* Animated orbs with responsive sizing and higher opacity */}
-      <div className="absolute inset-0">
-        <motion.div
-          className="absolute w-[50vw] md:w-[40vw] lg:w-[30vw] aspect-square bg-gradient-to-r from-purple-500/70 to-pink-500/70 rounded-full mix-blend-screen filter blur-xl"
-          animate={prefersReducedMotion ? {} : {
-            x: ['-10vw', '10vw', '-10vw'],
-            y: ['-10vh', '10vh', '-10vh'],
-            scale: [1, 1.2, 1],
-          }}
-          transition={{
-            duration: 20,
-            repeat: Infinity,
-            ease: "linear"
-          }}
-          style={{ top: '10%', left: '10%' }}
-        />
-        <motion.div
-          className="absolute w-[60vw] md:w-[45vw] lg:w-[35vw] aspect-square bg-gradient-to-r from-blue-500/70 to-cyan-500/70 rounded-full mix-blend-screen filter blur-xl"
-          animate={prefersReducedMotion ? {} : {
-            x: ['10vw', '-10vw', '10vw'],
-            y: ['10vh', '-10vh', '10vh'],
-            scale: [1, 1.3, 1],
-          }}
-          transition={{
-            duration: 25,
-            repeat: Infinity,
-            ease: "linear"
-          }}
-          style={{ top: '60%', right: '10%' }}
-        />
-        <motion.div
-          className="absolute w-[40vw] md:w-[30vw] lg:w-[25vw] aspect-square bg-gradient-to-r from-orange-500/70 to-yellow-500/70 rounded-full mix-blend-screen filter blur-xl"
-          animate={prefersReducedMotion ? {} : {
-            x: ['0vw', '15vw', '0vw'],
-            y: ['0vh', '-15vh', '0vh'],
-            scale: [1, 1.1, 1],
-          }}
-          transition={{
-            duration: 30,
-            repeat: Infinity,
-            ease: "linear"
-          }}
-          style={{ bottom: '10%', left: '30%' }}
-        />
-      </div>
-
-      {/* Grid overlay with responsive sizing */}
-      <div className="absolute inset-0 bg-[linear-gradient(rgba(255,255,255,0.03)_1px,transparent_1px),linear-gradient(90deg,rgba(255,255,255,0.03)_1px,transparent_1px)] bg-[size:25px_25px] sm:bg-[size:35px_35px] md:bg-[size:50px_50px]"></div>
-    </div>
-  );
-};
-
-// Loading Component
-const LoadingScreen = () => (
-  <motion.div
-    className="fixed inset-0 z-50 flex items-center justify-center"
-    style={{
-      background: 'linear-gradient(135deg, rgba(139,92,246,0.75) 0%, rgba(236,72,153,0.75) 100%)',
-      // This is a purple-to-pink gradient with 75% opacity
-      backdropFilter: 'blur(4px)',
-    }}
-    initial={{ opacity: 1 }}
-    exit={{ opacity: 0 }}
-    transition={{ duration: 0.5 }}
-  >
-    <div className="text-center">
-      <motion.div
-        className="w-16 h-16 mx-auto mb-4 rounded-full border-4 border-t-transparent"
-        style={{
-          borderColor: 'rgba(236,72,153,0.8) rgba(139,92,246,0.8) rgba(139,92,246,0.8) rgba(236,72,153,0.8)',
-          borderTopColor: 'transparent',
-        }}
-        animate={{ rotate: 360 }}
-        transition={{ duration: 1, repeat: Infinity, ease: "linear" }}
-      />
-      <motion.p
-        className="text-lg font-medium bg-gradient-to-r from-purple-400 via-pink-400 to-orange-400 bg-clip-text text-transparent"
-        initial={{ opacity: 0.5 }}
-        animate={{ opacity: 1 }}
-        transition={{ duration: 1, repeat: Infinity, repeatType: "reverse" }}
-      >
-        Loading Portfolio...
-      </motion.p>
-    </div>
-  </motion.div>
-);
-
-// Navbar Component
-const Navbar = () => {
-  const [scrolled, setScrolled] = useState(false);
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const [activeSection, setActiveSection] = useState('home');
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setScrolled(window.scrollY > 50);
-      
-      // Update active section based on scroll position
-      const sections = ['home', 'about', 'services', 'expertise', 'projects', 'pricing', 'contact'];
-      const scrollPosition = window.scrollY + window.innerHeight / 3;
-      
-      for (const section of sections) {
-        const element = document.getElementById(section);
-        if (element && scrollPosition >= element.offsetTop && 
-            scrollPosition < element.offsetTop + element.offsetHeight) {
-          setActiveSection(section);
-          break;
-        }
-      }
-    };
-    
-    window.addEventListener('scroll', handleScroll, { passive: true });
-    handleScroll(); // Initial check
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-
-  const navItems = [
-    { name: 'Home', href: '#home' },
-    { name: 'About', href: '#about' },
-    { name: 'Services', href: '#services' },
-    { name: 'Expertise', href: '#expertise' },
-    { name: 'Projects', href: '#projects' },
-    { name: 'Pricing', href: '#pricing' },
-    { name: 'Contact', href: '#contact' }
-  ];
-
-  const scrollToSection = (e, href) => {
-    e.preventDefault();
-    const element = document.querySelector(href);
-    if (element) {
-      const yOffset = -80;
-      const y = element.getBoundingClientRect().top + window.pageYOffset + yOffset;
-      window.scrollTo({ top: y, behavior: 'smooth' });
-    }
-    setMobileMenuOpen(false);
-  };
-
-  return (
-    <motion.nav
-      className={`fixed top-0 left-0 right-0 z-50 transition-all duration-300 ${
-        scrolled ? 'bg-slate-900/80 backdrop-blur-lg border-b border-slate-800/50' : 'bg-transparent'
-      }`}
-      initial={{ y: -100 }}
-      animate={{ y: 0 }}
-      transition={{ duration: 0.5 }}
-    >
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="flex items-center justify-between h-16">
-          <motion.a
-            href="#home"
-            className="font-bold text-xl sm:text-2xl bg-gradient-to-r from-purple-400 to-pink-400 bg-clip-text text-transparent"
-            whileHover={{ scale: 1.05 }}
-            onClick={(e) => scrollToSection(e, '#home')}
-          >
-            Aparna.dev
-          </motion.a>
-          
-          {/* Desktop Navigation */}
-          <div className="hidden md:flex items-center space-x-4 lg:space-x-8">
-            {navItems.map((item) => (
-              <motion.a
-                key={item.name}
-                href={item.href}
-                className={`text-sm lg:text-base transition-colors relative group ${
-                  activeSection === item.href.substring(1) 
-                    ? 'text-white font-medium' 
-                    : 'text-gray-300 hover:text-white'
-                }`}
-                whileHover={{ scale: 1.05 }}
-                onClick={(e) => scrollToSection(e, item.href)}
-              >
-                {item.name}
-                <span 
-                  className={`absolute -bottom-1 left-0 h-0.5 bg-purple-500 transition-all duration-300 ${
-                    activeSection === item.href.substring(1) ? 'w-full' : 'w-0 group-hover:w-full'
-                  }`}
-                ></span>
-              </motion.a>
-            ))}
-            <motion.a
-              href="#contact"
-              className="px-4 py-1.5 sm:px-6 sm:py-2 bg-gradient-to-r from-purple-500 to-pink-500 text-white text-sm lg:text-base rounded-full hover:shadow-lg hover:shadow-purple-500/30 transition-all"
-              whileHover={{ scale: 1.05 }}
-              whileTap={{ scale: 0.95 }}
-              onClick={(e) => scrollToSection(e, '#contact')}
-            >
-              Hire Me
-            </motion.a>
-          </div>
-          
-          {/* Mobile menu button */}
-          <div className="md:hidden">
-            <button
-              onClick={() => setMobileMenuOpen(!mobileMenuOpen)}
-              className="text-gray-300 hover:text-white p-2"
-              aria-label="Toggle menu"
-            >
-              {mobileMenuOpen ? <FiX size={24} /> : <FiMenu size={24} />}
-            </button>
-          </div>
-        </div>
-        
-        {/* Mobile Navigation */}
-        <AnimatePresence>
-          {mobileMenuOpen && (
-            <motion.div
-              className="md:hidden absolute top-16 left-0 right-0 bg-slate-900/90 backdrop-blur-lg border-b border-slate-800/50 shadow-lg"
-              initial={{ opacity: 0, height: 0 }}
-              animate={{ opacity: 1, height: 'auto' }}
-              exit={{ opacity: 0, height: 0 }}
-              transition={{ duration: 0.3 }}
-            >
-              <div className="px-4 py-4 space-y-4">
-                {navItems.map((item) => (
-                  <a
-                    key={item.name}
-                    href={item.href}
-                    className={`block transition-colors ${
-                      activeSection === item.href.substring(1)
-                        ? 'text-white font-medium'
-                        : 'text-gray-300 hover:text-white'
-                    }`}
-                    onClick={(e) => scrollToSection(e, item.href)}
-                  >
-                    {item.name}
-                  </a>
-                ))}
-                <a
-                  href="#contact"
-                  className="block w-full text-center px-6 py-2 bg-gradient-to-r from-purple-500 to-pink-500 text-white rounded-full"
-                  onClick={(e) => scrollToSection(e, '#contact')}
-                >
-                  Hire Me
-                </a>
-              </div>
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </div>
-    </motion.nav>
-  );
-};
-
-// Hero Section
-const HeroSection = () => {
-  return (
-    <section id="home" className="min-h-screen flex items-center justify-center pt-16">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="text-center">
-          <motion.div
-            initial={{ opacity: 0, y: 30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8 }}
-          >
-            <h1 className="text-4xl sm:text-5xl md:text-6xl lg:text-7xl font-bold mb-6">
-              <span className="bg-gradient-to-r from-purple-400 via-pink-400 to-orange-400 bg-clip-text text-transparent">
-                AI-Powered
-              </span>
-              <br />
-              <span className="text-white">Solutions</span>
-            </h1>
-            <p className="text-lg sm:text-xl md:text-2xl text-gray-300 mb-8 max-w-3xl mx-auto">
-              I build intelligent applications that transform businesses through cutting-edge AI integration and modern web technologies.
-            </p>
-          </motion.div>
-          <motion.div
-            className="flex flex-col sm:flex-row gap-4 justify-center items-center mb-12"
-            initial={{ opacity: 0, y: 30 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.8, delay: 0.2 }}
-          >
-            <motion.a
-              href="#contact"
-              className="px-8 py-4 bg-gradient-to-r from-purple-500 to-pink-500 text-white rounded-full text-lg font-semibold hover:shadow-lg hover:shadow-purple-500/30 transition-all"
-              whileHover={{ scale: 1.05 }}
-              whileTap={{ scale: 0.95 }}
-            >
-              Start Your Project
-            </motion.a>
-            <motion.a
-              href="#projects"
-              className="px-8 py-4 border-2 border-purple-500 text-purple-400 rounded-full text-lg font-semibold hover:bg-purple-500/10 transition-all"
-              whileHover={{ scale: 1.05 }}
-              whileTap={{ scale: 0.95 }}
-            >
-              View My Work
-            </motion.a>
-          </motion.div>
-          <motion.div
-            className="flex justify-center items-center space-x-8 text-gray-400"
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            transition={{ duration: 0.8, delay: 0.4 }}
-          >
-            <div className="flex items-center space-x-2">
-              <FaReact className="text-2xl text-cyan-400" />
-              <span>React</span>
-            </div>
-            <div className="flex items-center space-x-2">
-              <FaNode className="text-2xl text-green-400" />
-              <span>Node.js</span>
-            </div>
-            <div className="flex items-center space-x-2">
-              <FaPython className="text-2xl text-yellow-400" />
-              <span>Python</span>
-            </div>
-            <div className="flex items-center space-x-2">
-              <FaRobot className="text-2xl text-purple-400" />
-              <span>AI/ML</span>
-            </div>
-          </motion.div>
-        </div>
-      </div>
-    </section>
-  );
-};
-
-// About Section
-const AboutSection = () => {
-  return (
-    <section id="about" className="py-20 relative">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div className="grid lg:grid-cols-2 gap-12 items-center">
-          <motion.div
-            initial={{ opacity: 0, x: -50 }}
-            whileInView={{ opacity: 1, x: 0 }}
-            transition={{ duration: 0.8 }}
-            viewport={{ once: true }}
-          >
-            <SectionTitle
-              title="About Me"
-              subtitle="I'm Aparna Pradhan, a full-stack web and React Native developer deeply focused on building AI-integrated, niche-specific solutions — from automation and chatbots to SaaS tools and research systems."
-            />
-            <p className="text-lg text-gray-300 mb-6">
-              I&#39;m Aparna Pradhan, a full-stack web and React Native developer deeply focused on building AI-integrated, niche-specific solutions — from automation and chatbots to SaaS tools and research systems. I specialize in RAG, vector/graph DBs, LangChain, LangGraph, and memory-aware generation workflows.
-            </p>
-            <p className="text-lg text-gray-300 mb-8">
-              Based in West Bengal, India, I build modern, efficient AI-first applications that solve real-world problems for solopreneurs, SaaS startups, and small businesses. From agentic platforms to WhatsApp automation — I turn AI into leverage.
-            </p>
-            <div className="flex flex-wrap gap-4">
-              <span className="px-4 py-2 bg-purple-500/20 text-purple-300 rounded-full text-sm">AI + Fullstack Expert</span>
-              <span className="px-4 py-2 bg-pink-500/20 text-pink-300 rounded-full text-sm">Custom Agent Architect</span>
-              <span className="px-4 py-2 bg-orange-500/20 text-orange-300 rounded-full text-sm">Next.js | Expo | LangChain</span>
-            </div>
-          </motion.div>
-          <motion.div
-            initial={{ opacity: 0, x: 50 }}
-            whileInView={{ opacity: 1, x: 0 }}
-            transition={{ duration: 0.8 }}
-            viewport={{ once: true }}
-            className="relative"
-          >
-            <div className="w-full max-w-md mx-auto">
-              <div className="relative w-80 h-80 mx-auto">
-                <div className="absolute inset-0 bg-gradient-to-r from-purple-500 to-pink-500 rounded-full blur-lg opacity-30"></div>
-                <div className="relative w-full h-full bg-gradient-to-r from-purple-500/20 to-pink-500/20 rounded-full flex items-center justify-center border border-purple-500/30">
-                  <Image
-                    src={me}
-                    alt="Aparna Pradhan"
-                    fill
-                    className="object-cover rounded-full shadow-lg"
-                    priority
-                  />
-                </div>
-              </div>
-            </div>
-          </motion.div>
-        </div>
-      </div>
-    </section>
-  );
-};
-
-// Services Section
-const ServicesSection = () => {
-  const services = [
-    {
-      icon: <FiCode className="w-8 h-8" />,
-      title: "Full-Stack Development",
-      description: "End-to-end web applications with modern technologies like React, Next.js, and Node.js",
-      color: "from-blue-500 to-cyan-500"
-    },
-    {
-      icon: <FaRobot className="w-8 h-8" />,
-      title: "AI Integration",
-      description: "Seamlessly integrate AI capabilities into your applications with LLMs, machine learning, and automation",
-      color: "from-purple-500 to-pink-500"
-    },
-    {
-      icon: <FaMobileAlt className="w-8 h-8" />,
-      title: "Mobile Development",
-      description: "Cross-platform mobile applications with React Native and modern mobile technologies",
-      color: "from-green-500 to-emerald-500"
-    },
-    {
-      icon: <FiServer className="w-8 h-8" />,
-      title: "Backend Development",
-      description: "Scalable server-side solutions with APIs, databases, and cloud infrastructure",
-      color: "from-orange-500 to-red-500"
-    },
-    {
-      icon: <FaBrain className="w-8 h-8" />,
-      title: "AI Consulting",
-      description: "Strategic guidance on AI adoption and implementation for your business needs",
-      color: "from-violet-500 to-purple-500"
-    },
-    {
-      icon: <FiDatabase className="w-8 h-8" />,
-      title: "Data Solutions",
-      description: "Database design, data analytics, and business intelligence solutions",
-      color: "from-teal-500 to-blue-500"
-    }
-  ];
-
-  return (
-    <section id="services" className="py-20 relative">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <motion.div
-          className="text-center mb-16"
-          initial={{ opacity: 0, y: 30 }}
-          whileInView={{ opacity: 1, y: 0 }}
-          transition={{ duration: 0.8 }}
-          viewport={{ once: true }}
-        >
-          <h2 className="text-4xl font-bold text-white mb-4">
-            My <span className="bg-gradient-to-r from-purple-400 to-pink-400 bg-clip-text text-transparent">Services</span>
-          </h2>
-          <p className="text-xl text-gray-300 max-w-3xl mx-auto">
-            I offer comprehensive solutions to help your business leverage modern technology and AI
-          </p>
-        </motion.div>
-        <div className="grid md:grid-cols-2 lg:grid-cols-3 gap-8">
-          {services.map((service, index) => (
-            <motion.div
-              key={index}
-              className="relative group"
-              initial={{ opacity: 0, y: 30 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.8, delay: index * 0.1 }}
-              viewport={{ once: true }}
-              whileHover={{ y: -10 }}
-            >
-              <div className="h-full p-8 bg-slate-800/50 backdrop-blur-sm rounded-xl border border-slate-700/50 hover:border-purple-500/50 transition-all duration-300">
-                <div className={`w-16 h-16 rounded-lg bg-gradient-to-r ${service.color} flex items-center justify-center mb-6 text-white group-hover:scale-110 transition-transform`}>
-                  {service.icon}
-                </div>
-                <h3 className="text-xl font-bold text-white mb-4">{service.title}</h3>
-                <p className="text-gray-300 leading-relaxed">{service.description}</p>
-              </div>
-            </motion.div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-};
-
-// Expertise Section from the second snippet with enhancements
-const ExpertiseSection = () => {
-  const expertise = [
-    {
-      title: "Full-Stack Development",
-      description: "End-to-end web development with modern technologies and best practices.",
-      color: "from-orange-900/30 to-orange-800/10",
-      icon: <FiCode className="w-6 h-6 text-orange-400" />,
-      techs: ["Next.js", "React", "Node.js", "MongoDB", "PostgreSQL", "Prisma", "Redis", "TailwindCSS", "ShadcnUI", "Stripe", "NextAuth", "JWT", "Clerk", "Upstash", "graphql"]
-    },
-    {
-      title: "Mobile Development",
-      description: "Building cross-platform mobile applications with React Native.",
-      color: "from-blue-900/30 to-blue-800/10",
-      icon: <FaMobileAlt className="w-6 h-6 text-blue-400" />,
-      techs: ["Expo", "nativewind", "zustand", "Firebase / appwrite / supabase / custom", "Push Notifications", "Offline First"]
-    },
-    {
-      title: "AI Integration",
-      description: "Seamlessly integrating AI capabilities into existing applications.",
-      color: "from-green-900/30 to-green-800/10",
-      icon: <FaRobot className="w-6 h-6 text-green-400" />,
-      techs: ["LLM API", "LangChain", "Vector DBs", "third party APIs", "AI agents", "workflow automation"]
-    },
-    {
-      title: "Backend Development",
-      description: "Scalable and efficient server-side solutions for your applications.",
-      color: "from-red-900/30 to-red-800/10",
-      icon: <FiServer className="w-6 h-6 text-red-400" />,
-      techs: ["Node.js", "Express", "Django", "FastAPI", "MongoDB", "PostgreSQL"]
-    }
-  ];
-
-  return (
-    <section id="expertise" className="py-20 relative">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <SectionTitle
-          title="My Expertise"
-          subtitle="I specialize in creating intelligent solutions that drive growth and efficiency for your business."
-        />
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-          {expertise.map((item, index) => (
-            <motion.div
-              key={index}
-              className={`bg-gradient-to-br ${item.color} rounded-2xl p-6 border border-orange-500/20 hover:border-orange-400/40 transition-all duration-300 group`}
-              whileHover={{ y: -5, scale: 1.02 }}
-              initial={{ opacity: 0, y: 20 }}
-              whileInView={{ opacity: 1, y: 0 }}
-              transition={{ duration: 0.5, delay: index * 0.1 }}
-              viewport={{ once: true }}
-            >
-              <div className="w-12 h-12 rounded-xl flex items-center justify-center mb-4">
-                {item.icon}
-              </div>
-              <h3 className={`text-xl font-bold text-white mb-3 ${spaceGrotesk.className}`}>{item.title}</h3>
-              <p className="text-gray-400 mb-4">{item.description}</p>
-              <div className="flex flex-wrap gap-2">
-                {item.techs.map((tech, i) => (
-                  <span key={i} className="px-3 py-1 bg-zinc-800/50 text-gray-200 text-xs rounded-full">
-                    {tech}
-                  </span>
-                ))}
-              </div>
-            </motion.div>
-          ))}
-        </div>
-      </div>
-    </section>
-  );
-};
-
-// Utility section component from the second snippet
+// Utility section component from the second snippet - (Not used in ModernPortfolio render)
 const Section = ({ id, title, children, subtitle }) => (
   <section id={id} className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-16 sm:py-20">
     <SectionTitle title={title} subtitle={subtitle} />
@@ -921,94 +379,8 @@ const PricingSection = () => {
   );
 };
 
-// Contact Section
-const ContactSection = () => {
-  return (
-    <section id="contact" className="relative pt-20 pb-32 sm:pt-24 sm:pb-40 overflow-hidden">
-      <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 relative">
-        <div className="text-center mb-12">
-          <h2 className={`text-3xl sm:text-4xl font-bold text-white mb-4 ${spaceGrotesk.className}`}>
-            Ready to Transform Your Business with AI?
-          </h2>
-          <p className={`text-lg text-gray-300 max-w-2xl mx-auto ${firaCode.className}`}>
-            I&#39;m here to help you navigate the world of AI and build solutions that make an impact.
-          </p>
-        </div>
-
-        <div className="bg-gradient-to-br from-gray-900/90 to-gray-800/60 backdrop-blur-sm rounded-2xl overflow-hidden border border-gray-700/50 p-6 sm:p-8 lg:p-10 shadow-xl">
-          <div className="grid md:grid-cols-2 gap-8 items-center">
-            <div>
-              <h3 className={`text-2xl font-bold text-white mb-4 ${spaceGrotesk.className}`}>Get in Touch</h3>
-              <p className="text-gray-300 mb-6">
-                Have a project in mind or want to discuss how AI can benefit your business?
-                I&#39;m here to help you navigate the world of AI and build solutions that make an impact.
-              </p>
-
-              <div className="space-y-4 mb-6">
-                <a
-                  href="mailto:softservicesinc.portfolio@gmail.com"
-                  className="flex items-center text-gray-300 hover:text-white transition-colors"
-                >
-                  <FiMail className="w-5 h-5 mr-3 text-orange-400" />
-                  softservicesinc.portfolio@gmail.com
-                </a>
-                <a
-                  href="https://goo.gl/maps/SQUjHtzSMfeZfmWR7"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="flex items-center text-gray-300 hover:text-white transition-colors"
-                >
-                  <FiMapPin className="w-5 h-5 mr-3 text-orange-400" />
-                  West Bengal, India
-                </a>
-              </div>
-
-              <div className="mt-8">
-                <h4 className="text-lg font-semibold text-white mb-3">Connect with me</h4>
-                <div className="flex space-x-4">
-                  <a
-                    href="https://www.linkedin.com/in/aparna-pradhan-06b882215/"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="w-10 h-10 rounded-full bg-gray-800 hover:bg-blue-600 flex items-center justify-center text-gray-300 hover:text-white transition-colors"
-                    aria-label="LinkedIn"
-                  >
-                    <FiLinkedin className="w-5 h-5" />
-                  </a>
-                  <a
-                    href="https://x.com/Aparna_108_dev/"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    className="w-10 h-10 rounded-full bg-gray-800 hover:bg-blue-400 flex items-center justify-center text-gray-300 hover:text-white transition-colors"
-                    aria-label="Twitter"
-                  >
-                    <FiTwitter className="w-5 h-5" />
-                  </a>
-                </div>
-              </div>
-            </div>
-
-            <div className="bg-gradient-to-br from-purple-900/40 to-orange-900/30 p-6 sm:p-8 rounded-xl border border-purple-500/30 h-full flex flex-col justify-center backdrop-blur-sm shadow-lg">
-              <h4 className={`text-xl font-bold text-white mb-4 ${spaceGrotesk.className}`}>Send Me a Message</h4>
-              <p className="text-gray-300 mb-6">
-                Have a project in mind or questions about my services? Feel free to reach out through email or any of my social media channels.
-              </p>
-              <a
-                href="mailto:softservicesinc.portfolio@gmail.com"
-                className="inline-flex items-center justify-center w-full px-6 py-3 text-lg font-medium text-white bg-gradient-to-r from-orange-500 to-pink-500 rounded-lg hover:from-orange-600 hover:to-pink-600 transition-all shadow-lg hover:shadow-orange-500/30"
-              >
-                <FiMail className="w-5 h-5 mr-2" />
-                Send Email
-              </a>
-            </div>
-          </div>
-        </div>
-      </div>
-    </section>
-  );
-};
-
 // Import enhanced UI components
+import ContactSection from './component/ContactSection'; // Added import
 import ProgressHeader from './component/navigation/ProgressHeader';
 import FloatingNavigation from './component/ui/FloatingNavigation';
 import ScrollProgress from './component/ui/ScrollProgress';
